# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TT Stock Backend API
  description: |
    A comprehensive tire and wheel inventory management system API built with Go and Gin framework.
    
    ## Features
    - User authentication and authorization with JWT tokens
    - Product management (tires and wheels)
    - Stock movement tracking
    - Low stock alerts
    - Business intelligence and reporting
    - Role-based access control (Admin, Owner, Staff)
    
    ## Authentication
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to prevent abuse. Current limits:
    - 100 requests per minute per IP
    - 1000 requests per hour per user
    
    ## Response Times
    All API endpoints are designed to respond within 200ms under normal load conditions.
  version: 1.0.0
  contact:
    name: TT Stock API Support
    email: support@ttstock.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ttstock.com
    description: Production server
...
paths:
...
  /api/v1/alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Get a paginated list of alerts
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: type
          in: query
          description: Filter by alert type
          required: false
          schema:
            type: string
            enum: [low_stock, out_of_stock, price_change, system]
        - name: priority
          in: query
          description: Filter by priority
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: unread_only
          in: query
          description: Show only unread alerts
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/alerts/{id}/read:
    put:
      tags:
        - Alerts
      summary: Mark alert as read
      description: Mark an alert as read
      operationId: markAlertAsRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Alert ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AlertResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: [low_stock, out_of_stock, price_change, system]
          example: "low_stock"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "medium"
        title:
          type: string
          example: "Low Stock Alert"
        message:
          type: string
          example: "Product MIC-PS4-225-45-17 is running low on stock"
        product_id:
          type: integer
          format: int64
          nullable: true
          example: 1
        is_read:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    AlertListResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/AlertResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request data"
        message:
          type: string
          example: "The provided data is invalid"
        details:
          type: object
          additionalProperties: true
          example:
            field: "phone_number"
            reason: "Phone number is required"
security:
  - bearerAuth: []